Add files via upload
Added the player file, added ability for player to walk through the maze to a randomly generated exit. Right now the player just teleports back to spawn with the same maze, but I will try to get the program to generate a new maze instead. Basically I need to learn how to restart the program, probably through html.