commit all files necessary for website
Please note this isn't the finished website by any means, but this has all of the files necessary to run both games on the website. Please make sure your file paths are correct if you want to open up the page with the games on it. If you download the files as they are in the website folder, the website will function properly.